+++ d2/src/main/java/com/linkedin/d2/balancer/D2ClientConfig.java
@@ -37,9 +37,9 @@ import com.linkedin.d2.discovery.event.ServiceDiscoveryEventEmitter;
 import com.linkedin.d2.discovery.stores.zk.ZKPersistentConnection;
 import com.linkedin.d2.discovery.stores.zk.ZooKeeper;
 import com.linkedin.d2.discovery.stores.zk.ZooKeeperStore;
-import com.linkedin.d2.jmx.IndisObserverMetricsProvider;
+import com.linkedin.d2.jmx.XdsServerMetricsProvider;
 import com.linkedin.d2.jmx.JmxManager;
-import com.linkedin.d2.jmx.NoOpIndisObserverMetricsProvider;
+import com.linkedin.d2.jmx.NoOpXdsServerMetricsProvider;
 import com.linkedin.d2.jmx.NoOpJmxManager;
 import com.linkedin.r2.transport.common.TransportClientFactory;
 import io.grpc.netty.shaded.io.netty.handler.ssl.SslContext;
@@ -137,7 +137,7 @@ public class D2ClientConfig
   public String xdsChannelLoadBalancingPolicy = null;
   public Map<String, ?> xdsChannelLoadBalancingPolicyConfig = null;
   public boolean subscribeToUriGlobCollection = false;
-  public IndisObserverMetricsProvider indisObserverMetricsProvider = new NoOpIndisObserverMetricsProvider();
+  public XdsServerMetricsProvider _xdsServerMetricsProvider = new NoOpXdsServerMetricsProvider();
 
   public D2ClientConfig()
   {
@@ -214,7 +214,7 @@ public class D2ClientConfig
                  String xdsChannelLoadBalancingPolicy,
                  Map<String, ?> xdsChannelLoadBalancingPolicyConfig,
                  boolean subscribeToUriGlobCollection,
-                 IndisObserverMetricsProvider indisObserverMetricsProvider
+                 XdsServerMetricsProvider xdsServerMetricsProvider
       )
   {
     this.zkHosts = zkHosts;
@@ -288,6 +288,6 @@ public class D2ClientConfig
     this.xdsChannelLoadBalancingPolicy = xdsChannelLoadBalancingPolicy;
     this.xdsChannelLoadBalancingPolicyConfig = xdsChannelLoadBalancingPolicyConfig;
     this.subscribeToUriGlobCollection = subscribeToUriGlobCollection;
-    this.indisObserverMetricsProvider = indisObserverMetricsProvider;
+    this._xdsServerMetricsProvider = xdsServerMetricsProvider;
   }
 }