+++ d2/src/main/java/com/linkedin/d2/xds/balancer/XdsLoadBalancerWithFacilitiesFactory.java
@@ -58,7 +58,7 @@ public class XdsLoadBalancerWithFacilitiesFactory implements LoadBalancerWithFac
         executorService,
         xdsStreamReadyTimeout,
         config.subscribeToUriGlobCollection,
-        config.indisObserverMetricsProvider
+        config._xdsServerMetricsProvider
     );
     d2ClientJmxManager.registerXdsClientJmx(xdsClient.getXdsClientJmx());